move amountValidation to exchangeService
