+++ src/main/java/com/globant/model/Wallet.java
@@ -16,28 +16,20 @@ public class Wallet {
     }
 
     public void receiveFiat(BigDecimal amount) {
-        amountValidation(amount);
         this.fiatBalance = this.fiatBalance.add(amount);
     }
 
     public void deliverFiat(BigDecimal amount) {
-        amountValidation(amount);
         fundsValidation(amount, this.fiatBalance);
         this.fiatBalance = this.fiatBalance.subtract(amount);
     }
 
     public void receiveCrypto(String crypto, BigDecimal amount) {
-        amountValidation(amount);
         BigDecimal currentBalance = this.cryptoBalances.getOrDefault(crypto, BigDecimal.ZERO);
         BigDecimal newBalance = currentBalance.add(amount);
         this.cryptoBalances.put(crypto, newBalance);
     }
 
-    private void amountValidation(BigDecimal amount) {
-        if (amount.compareTo(BigDecimal.ZERO) <= 0) {
-            throw new NegativeAmountException();
-        }
-    }
 
     private void fundsValidation(BigDecimal amount, BigDecimal funds) {
         if (amount.compareTo(funds) > 0) {
